chore: rename shared components
