+++ app/admin/layout.tsx
@@ -1,8 +1,9 @@
 import type { Metadata } from 'next'
 
-import { HeaderComponent } from '@/components/shared/header-component'
-import { Footercomponent } from '@/components/shared/footer-component'
-import { FloatingMenuComponent } from '@/components/shared/floating-menu-component'
+import { Header } from '@/components/shared/header'
+import { Footer } from '@/components/shared/footer'
+
+import { FloatingMenu } from '@/components/shared/floating-menu'
 
 import { Toaster } from '@/components/ui/toaster'
 
@@ -15,10 +16,10 @@ const Layout = ({
 }>) => {
   return (
     <div className='flex flex-col w-full min-h-screen'>
-      <HeaderComponent />
+      <Header />
       {children}
-      <Footercomponent />
-      <FloatingMenuComponent />
+      <Footer />
+      <FloatingMenu />
       <Toaster />
     </div>
   )
